--- lib/config/constants.dart
@@ -1,10 +1,0 @@
-class AppConstants {
-  /// URL to Terms & Conditions
-  static const String tncURL = "https://google.com"; //TODO
-
-  /// URL to Privacy Policy
-  static const String privacyURL = "https://google.com"; //TODO
-
-  // API URLs
-  static const String registerURL = "http://bitecope.co.in:8080/register/";
-}
