+++ lib/config/constants/app_urls.dart
@@ -1,0 +1,11 @@
+class AppURLs {
+  /// URL to Terms & Conditions
+  static const String tnc = "https://google.com"; //TODO
+
+  /// URL to Privacy Policy
+  static const String privacy = "https://google.com"; //TODO
+
+  // API URLs
+  static const String baseAPI = "http://bitecope.co.in:8080/";
+  static const String register = "$baseAPI/register";
+}
