+++ lib/config/routes/router.dart
@@ -1,8 +1,8 @@
-import 'package:bitecope/data/repositories/account_repository.dart';
-import 'package:bitecope/logic/sign_up/sign_up_bloc.dart';
-import 'package:bitecope/screens/sign_up/pages/sign_up_one.dart';
-import 'package:bitecope/screens/splash_screen/splash_screen.dart';
-import 'package:bitecope/screens/not_found/not_found.dart';
+import 'package:bitecope/modules/sign_up/repositories/sign_up_repository.dart';
+import 'package:bitecope/modules/not_found/screens/not_found.dart';
+import 'package:bitecope/modules/sign_up/bloc/sign_up_bloc.dart';
+import 'package:bitecope/modules/sign_up/pages/sign_up_one.dart';
+import 'package:bitecope/modules/splash_screen/screens/splash_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -20,7 +20,7 @@ class AppRouter {
           builder: (_) {
             return BlocProvider<SignUpBloc>(
               create: (context) => SignUpBloc(
-                accountRepository: AccountRepository(),
+                accountRepository: SignUpRepository(),
               ),
               child: const SignUpOne(),
             );
