+++ lib/core/authentication/bloc/authentication_state.dart
