+++ lib/core/authentication/models/user.dart
