+++ lib/modules/sign_up/models/sign_up_response.dart
